update to remove dependency on perl for 'list' feature
The install relies on perl for a single-line rexex when scraping the
Bintray website for a list of REX-Ray versions available for install.

This tiny update removes the need for the installing operating system to
have Perl installed, mostly because RHEL does not have Perl installed by
default. One other option would have been to trap the error and echo
some sort of message calling for the user to install Perl, but that
seems drastic, considering the same page scrape can be executed just
fine using the bash built-in command 'cut'.

This update should not affect any other versions, as 'cut' hasn't really
changed its syntax in decades.