Updated Vagrantfile
This patch updates the Vagrantfile to ensure all URLs and file paths are
escaped for use by the shell interpreter. Additionally, the
`$volume_path` variable is no longer dependent upon the person executing
the `Hello REX-Ray` example to customize it. Instead the `$volume_path`
variable is simply set to the path `$(pwd)/.vagrant/volumes`. Finally,
the VirtualBox web server is automatically started at the head of the
Vagrantfile if it is not already running.